Flatten packages/lesswrong/client
