+++ packages/lesswrong/client/apolloClient.ts
@@ -1,3 +1,6 @@
+import { ApolloClient, InMemoryCache, ApolloLink } from '@apollo/client';
+import { BatchHttpLink } from '@apollo/client/link/batch-http';
+import { apolloCacheVoteablePossibleTypes } from '../lib/make_voteable';
 import { onError } from '@apollo/client/link/error';
 import type { SourceLocation } from 'graphql';
 
@@ -16,4 +19,22 @@ const errorLink = onError(({ graphQLErrors, networkError }) => {
   }
 });
 
-export default errorLink;
+export const createApolloClient = () => {
+  const cache = new InMemoryCache({
+    possibleTypes: {
+      ...apolloCacheVoteablePossibleTypes()
+    }
+  })
+    .restore((window as any).__APOLLO_STATE__); //ssr
+  
+  const httpLink = new BatchHttpLink({
+    uri: '/graphql',
+    credentials: 'same-origin',
+    batchMax: 50,
+  });
+  
+  return new ApolloClient({
+    link: ApolloLink.from([errorLink, httpLink]),
+    cache
+  });
+};
