--- packages/lesswrong/client/vulcan-lib/main.ts
@@ -1,3 +1,0 @@
-import './apollo-client/apolloClient';
-import './apollo-client/updates';
-import './apollo-client/links/error';
